+++ src/main/java/http/controller/Controller.java
@@ -1,11 +1,11 @@
-package http.path;
+package http.controller;
 
 import http.RequestBody;
 import http.RequestHeader;
 
 import java.io.DataOutputStream;
 
-public interface Path {
+public interface Controller {
     void get(DataOutputStream dos, RequestHeader requestHeader);
     void post(DataOutputStream dos, RequestHeader requestHeader, RequestBody requestBody);
 }
