+++ src/main/java/webserver/RequestHandler.java
@@ -1,10 +1,13 @@
 package webserver;
 
-import http.*;
-import http.path.Index;
-import http.path.Path;
-import http.path.RawFile;
-import http.path.UserCreate;
+import http.RequestBody;
+import http.RequestHeader;
+import http.RequestLine;
+import http.ResponseHeader;
+import http.controller.Controller;
+import http.controller.IndexController;
+import http.controller.RawFileController;
+import http.controller.UserCreateController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -36,20 +39,20 @@ public class RequestHandler implements Runnable {
     }
 
     private void route(BufferedReader br, DataOutputStream dos) throws IOException {
-        Map<String, Path> paths = new HashMap<>();
-        paths.put("/user/create", new UserCreate());
-        paths.put("/", new Index());
+        Map<String, Controller> controllers = new HashMap<>();
+        controllers.put("/user/create", new UserCreateController());
+        controllers.put("/", new IndexController());
         try {
             RequestLine requestLine = new RequestLine(br);
             RequestHeader requestHeader = new RequestHeader(br);
 
-            Path path = paths.getOrDefault(requestLine.getPath(), new RawFile(requestLine.getPath()));
+            Controller controller = controllers.getOrDefault(requestLine.getPath(), new RawFileController(requestLine.getPath()));
 
             if (requestLine.isMethodEqualTo("GET")) {
-                path.get(dos, requestHeader);
+                controller.get(dos, requestHeader);
             } else if (requestLine.isMethodEqualTo("POST")) {
                 RequestBody requestBody = new RequestBody(br, requestHeader.getContentLength());
-                path.post(dos, requestHeader, requestBody);
+                controller.post(dos, requestHeader, requestBody);
             }
         } catch (IllegalArgumentException e) {
             ResponseHeader.response400Header(dos);
